Move everything related to registries from InternalPacketListener to new class
